todays work

+++ ERP/_Estimate.cs
@@ -12,9 +12,9 @@ namespace ERP
         public int Cust_ID { get; set; }
         public int Employee_ID {get;set;}
         public string Estimate_Date { get; set; }
-        public long Estimate_Subtotal { get; set; }
-        public long Estimate_Tax { get; set; }
-        public long Estimate_Total { get; set; }
+        public double Estimate_Subtotal { get; set; }
+        public double Estimate_Tax { get; set; }
+        public double Estimate_Total { get; set; }
         public string Estimate_BillStreet { get; set; }
         public string Estimate_BillCity { get; set; }
         public string Estimate_BillState { get; set; }
